--- src/__tests__/__mocks__/fetchStocks.js
@@ -1,17 +1,0 @@
-const listOfStocks = [
-  {
-    symbol: 'INTC',
-    name: 'Intel Corporation',
-    change: -0.06,
-    price: 48.01,
-    changesPercentage: -0.124820836,
-  }, {
-    symbol: 'MSFT',
-    name: 'Microsoft Corporation',
-    change: -7.655,
-    price: 282.205,
-    changesPercentage: -2.6409297,
-  },
-];
-
-export default listOfStocks;
